+++ packages/excalidraw/clipboard.ts
@@ -2,7 +2,7 @@ import type {
   ExcalidrawElement,
   NonDeletedExcalidrawElement,
 } from "./element/types";
-import type { BinaryFiles } from "./types";
+import type { AppState, BinaryFiles } from "./types";
 import type { Spreadsheet } from "./charts";
 import { tryParseSpreadsheet, VALID_SPREADSHEET } from "./charts";
 import {
@@ -333,6 +333,7 @@ const parseClipboardEvent = async (
 export const parseClipboard = async (
   event: ClipboardEvent,
   isPlainPaste = false,
+  appState?: AppState,
 ): Promise<ClipboardData> => {
   const parsedEventData = await parseClipboardEvent(event, isPlainPaste);
 
@@ -349,6 +350,10 @@ export const parseClipboard = async (
       !isPlainPaste && parsePotentialSpreadsheet(parsedEventData.value);
 
     if (spreadsheetResult) {
+      if ("spreadsheet" in spreadsheetResult) {
+        spreadsheetResult.spreadsheet.activeSubtypes = appState?.activeSubtypes;
+        spreadsheetResult.spreadsheet.customData = appState?.customData;
+      }
       return spreadsheetResult;
     }
   } catch (error: any) {