+++ packages/excalidraw/element/textWysiwyg.tsx
@@ -25,6 +25,7 @@ import {
   getContainerElement,
   getTextElementAngle,
   getTextWidth,
+  measureText,
   normalizeText,
   redrawTextBoundingBox,
   wrapText,
@@ -46,12 +47,15 @@ import {
 import type App from "../components/App";
 import { LinearElementEditor } from "./linearElementEditor";
 import { parseClipboard } from "../clipboard";
+import type { SubtypeMethods } from "./subtypes";
+import { getSubtypeMethods } from "./subtypes";
 import {
   originalContainerCache,
   updateOriginalContainerCache,
 } from "./containerCache";
 
 const getTransform = (
+  offsetX: number,
   width: number,
   height: number,
   angle: number,
@@ -69,8 +73,17 @@ const getTransform = (
   if (height > maxHeight && zoom.value !== 1) {
     translateY = (maxHeight * (zoom.value - 1)) / 2;
   }
-  return `translate(${translateX}px, ${translateY}px) scale(${zoom.value}) rotate(${degree}deg)`;
+  const offset = offsetX !== 0 ? ` translate(${offsetX}px, 0px)` : "";
+  return `translate(${translateX}px, ${translateY}px) scale(${zoom.value}) rotate(${degree}deg)${offset}`;
 };
+
+const getEditorStyle = function (element) {
+  const map = getSubtypeMethods(element.subtype);
+  if (map?.getEditorStyle) {
+    return map.getEditorStyle(element);
+  }
+  return {};
+} as SubtypeMethods["getEditorStyle"];
 
 export const textWysiwyg = ({
   id,
@@ -137,14 +150,27 @@ export const textWysiwyg = ({
         app.scene.getNonDeletedElementsMap(),
       );
 
-      let width = updatedTextElement.width;
+      // Editing metrics
+      const eMetrics = measureText(
+        container && updatedTextElement.containerId
+          ? wrapText(
+              updatedTextElement.originalText,
+              getFontString(updatedTextElement),
+              getBoundTextMaxWidth(container, updatedTextElement),
+            )
+          : updatedTextElement.originalText,
+        getFontString(updatedTextElement),
+        updatedTextElement.lineHeight,
+      );
 
-      // set to element height by default since that's
+      let width = Math.max(updatedTextElement.width, eMetrics.width);
+
+      // Set to element height by default since that's
       // what is going to be used for unbounded text
-      let height = updatedTextElement.height;
+      let height = Math.max(updatedTextElement.height, eMetrics.height);
 
-      let maxWidth = updatedTextElement.width;
-      let maxHeight = updatedTextElement.height;
+      let maxWidth = width;
+      let maxHeight = height;
 
       if (container && updatedTextElement.containerId) {
         if (isArrowElement(container)) {
@@ -240,8 +266,31 @@ export const textWysiwyg = ({
         width += 0.5;
       }
 
+      // Horizontal offset in case updatedTextElement has a non-WYSIWYG subtype
+      const offWidth = container
+        ? Math.min(
+            0,
+            updatedTextElement.width - Math.min(maxWidth, eMetrics.width),
+          )
+        : Math.min(maxWidth, updatedTextElement.width) -
+          Math.min(maxWidth, eMetrics.width);
+      const offsetX =
+        textAlign === "right"
+          ? offWidth
+          : textAlign === "center"
+          ? offWidth / 2
+          : 0;
+      let { width: w, height: h } = updatedTextElement;
+
       // add 5% buffer otherwise it causes wysiwyg to jump
       height *= 1.05;
+      h *= 1.05;
+
+      const transformOrigin =
+        updatedTextElement.width !== eMetrics.width ||
+        updatedTextElement.height !== eMetrics.height
+          ? { transformOrigin: `${w / 2}px ${h / 2}px` }
+          : {};
 
       const font = getFontString(updatedTextElement);
 
@@ -261,7 +310,9 @@ export const textWysiwyg = ({
         height: `${height}px`,
         left: `${viewportX - padding}px`,
         top: `${viewportY}px`,
+        ...transformOrigin,
         transform: getTransform(
+          offsetX,
           width,
           height,
           getTextElementAngle(updatedTextElement, container),
@@ -322,6 +373,7 @@ export const textWysiwyg = ({
     whiteSpace,
     overflowWrap: "break-word",
     boxSizing: "content-box",
+    ...getEditorStyle(element),
   });
   editable.value = element.originalText;
   updateWysiwygStyle();